Add enum definitions for various statuses and availability in the application
