+++ server/app/Enums/ShopsPartenerShip.php
