+++ server/app/Http/Controllers/Dashboard/IndexController.php
@@ -9,7 +9,6 @@ use App\Models\Product;
 use App\Models\Service;
 use App\Models\ServiceRequest;
 use App\Models\User;
-use App\OrdersStatus;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\DB;
